fix (init): change

+++ kernel/defs.h
@@ -109,7 +109,7 @@ void            ps(int pid);
 int             meminfo();
 int             waitpid(int pid);
 //TODO: Project03 Implement 3 syscalls
-int		mmap(uint64 addr, int length, int prot, int flags, int fd, int offset);
+uint64	mmap(uint64 addr, int length, int prot, int flags, int fd, int offset);
 int		munmap(uint64 addr);
 int		freemem();
