+++ user/user.h
@@ -33,7 +33,7 @@ int meminfo();
 int waitpid(int pid);
 
 //TODO: Project 03 Implement 3 syscalls
-int mmap(uint64 addr, int length, int prot, int flags, int fd, int offset);
+uint64 mmap(uint64 addr, int length, int prot, int flags, int fd, int offset);
 int munmap(uint64 addr);
 int freemem();
